Retire api.jsx and write test cases
